+++ service/controller/resource/servicemonitor/resource.go
@@ -1,9 +1,12 @@
 package servicemonitor
 
 import (
+	promv1 "github.com/coreos/prometheus-operator/pkg/apis/monitoring/v1"
 	promclient "github.com/coreos/prometheus-operator/pkg/client/versioned"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
+	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/servicemonitor/service"
+	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )
 
 const (
@@ -38,4 +41,15 @@ func New(config Config) (*Resource, error) {
 
 func (r *Resource) Name() string {
 	return Name
+}
+
+func toServiceMonitors(obj interface{}) ([]*promv1.ServiceMonitor, error) {
+	cluster, err := key.ToCluster(obj)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	return []*promv1.ServiceMonitor{
+		service.APIServer(cluster),
+	}, nil
 }
